UBUNTU: Packaging: Introduce copy-files and local-mangle
BugLink: https://bugs.launchpad.net/bugs/1817734

Two new scripts are introduced in order to do some of the copying and mangling
of copies that update-from-*master does.

One of the changes on copy-files compared to update-from-*master is that the -c
option is given to rsync, so it compares checksums of files in order to decide
whether they are different and need an update. That's necessary because
sometimes files will have the same size and their modified time will be whithin
one second or the original file will be older because git checked it out
earlier.

The script is split in two so the copy-files may be shared between different
kernel tress and the very specific changes are done on the local-mangle file,
which is different between trees.

Also, in order to make the copy-files the same one for all trees, some of the
copies and updates are dependent on a local.conf file, which is present only on
those trees where it's needed. The contents of those files are not so easily
generated, so they are not part of update.conf.

Signed-off-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>